Dataflow finalizado
